+++ lib/reteno.dart
@@ -2,6 +2,7 @@ library reteno;
 
 export './src/reteno.dart';
 export 'src/models/anonymous_user_attributes.dart';
+export 'src/models/in_app_message_status.dart';
 export 'src/models/reteno_custom_event.dart';
 export 'src/models/reteno_custom_event_parameter.dart';
 export 'src/models/reteno_user.dart';