+++ lib/src/reteno_plugin_pigeon_channel.dart
@@ -3,6 +3,7 @@ import 'dart:io';
 
 import 'package:reteno_plugin/src/extensions.dart';
 import 'package:reteno_plugin/src/models/anonymous_user_attributes.dart';
+import 'package:reteno_plugin/src/models/in_app_message_status.dart';
 import 'package:reteno_plugin/src/models/reteno_custom_event.dart';
 import 'package:reteno_plugin/src/models/reteno_user.dart';
 import 'package:reteno_plugin/src/reteno_plugin_platform_interface.dart';
@@ -10,6 +11,9 @@ import 'package:reteno_plugin/src/reteno_plugin_platform_interface.dart';
 import 'native_reteno_plugin.g.dart';
 
 typedef OnNotificationCallback = void Function(Map<String?, Object?> payload);
+typedef OnInAppMessageStatusChanged = void Function(
+  InAppMessageStatus status,
+);
 
 class RetenoPigeonChannel extends RetenoPluginPlatform {
   static RetenoPigeonChannel? _instance;
@@ -36,6 +40,9 @@ class RetenoPigeonChannel extends RetenoPluginPlatform {
         },
         onRetenoNotificationClicked: (payload) {
           onRetenoNotificationClicked.add(Map<String, dynamic>.from(payload));
+        },
+        onInAppMessageStatusChangedCallback: (status) {
+          onInAppMessageStatusChanged.add(status);
         },
       ),
     );
@@ -80,16 +87,23 @@ class RetenoPigeonChannel extends RetenoPluginPlatform {
     }
     return true;
   }
+
+  @override
+  Future<void> pauseInAppMessages(bool isPaused) {
+    return _api.pauseInAppMessages(isPaused);
+  }
 }
 
 class _RetenoFlutterApi extends RetenoFlutterApi {
   _RetenoFlutterApi({
     required this.onNotificationCallback,
     required this.onRetenoNotificationClicked,
+    required this.onInAppMessageStatusChangedCallback,
   });
 
   OnNotificationCallback onNotificationCallback;
   OnNotificationCallback onRetenoNotificationClicked;
+  OnInAppMessageStatusChanged onInAppMessageStatusChangedCallback;
 
   @override
   void onNotificationClicked(Map<String?, Object?> payload) =>
@@ -98,4 +112,47 @@ class _RetenoFlutterApi extends RetenoFlutterApi {
   @override
   void onNotificationReceived(Map<String?, Object?> payload) =>
       onNotificationCallback(payload);
+
+  @override
+  void onInAppMessageStatusChanged(
+    NativeInAppMessageStatus status,
+    NativeInAppMessageAction? action,
+    String? error,
+  ) =>
+      onInAppMessageStatusChangedCallback(
+          status.toInAppMessageStatus(action, error));
+}
+
+extension on NativeInAppMessageStatus {
+  InAppMessageStatus toInAppMessageStatus(
+      NativeInAppMessageAction? action, String? error) {
+    switch (this) {
+      case NativeInAppMessageStatus.inAppShouldBeDisplayed:
+        return InAppShouldBeDisplayed();
+      case NativeInAppMessageStatus.inAppIsDisplayed:
+        return InAppIsDisplayed();
+      case NativeInAppMessageStatus.inAppShouldBeClosed:
+        return InAppShouldBeClosed(
+          action: action!.toInAppMessageAction(),
+        );
+      case NativeInAppMessageStatus.inAppIsClosed:
+        return InAppIsClosed(
+          action: action!.toInAppMessageAction(),
+        );
+      case NativeInAppMessageStatus.inAppReceivedError:
+        return InAppReceivedError(
+          errorMessage: error!,
+        );
+    }
+  }
+}
+
+extension on NativeInAppMessageAction {
+  InAppMessageAction toInAppMessageAction() {
+    return InAppMessageAction(
+      isCloseButtonClicked: isCloseButtonClicked,
+      isButtonClicked: isButtonClicked,
+      isOpenUrlClicked: isOpenUrlClicked,
+    );
+  }
 }