+++ lib/src/reteno_plugin_platform_interface.dart
@@ -1,6 +1,7 @@
 import 'dart:async';
 
 import 'package:reteno_plugin/src/models/anonymous_user_attributes.dart';
+import 'package:reteno_plugin/src/models/in_app_message_status.dart';
 import 'package:reteno_plugin/src/models/reteno_custom_event.dart';
 import 'package:reteno_plugin/src/models/reteno_user.dart';
 
@@ -21,6 +22,9 @@ abstract class RetenoPluginPlatform {
   StreamController<Map<String, dynamic>> onRetenoNotificationClicked =
       StreamController<Map<String, dynamic>>.broadcast();
 
+  StreamController<InAppMessageStatus> onInAppMessageStatusChanged =
+      StreamController<InAppMessageStatus>.broadcast();
+
   Future<Map<dynamic, dynamic>?> getInitialNotification() {
     throw UnimplementedError(
         'getInitialNotification() has not been implemented.');
@@ -33,5 +37,9 @@ abstract class RetenoPluginPlatform {
   Future<bool> updatePushPermissionStatus() {
     throw UnimplementedError(
         'updatePushPermissionStatus() has not been implemented.');
+  }
+
+  Future<void> pauseInAppMessages(bool isPaused) {
+    throw UnimplementedError('pauseInAppMessages() has not been implemented.');
   }
 }